+++ lib/puppet/type/wls_authentication_provider.rb
@@ -42,11 +42,13 @@ module Puppet
     parameter :domain
     parameter :name
     parameter :authentication_provider_name
+    parameter :timeout
 
     property :control_flag
     parameter :providerclassname
     parameter :attributes
     parameter :attributesvalues
+    parameter :order
 
     add_title_attributes(:authentication_provider_name) do
       /^((.*\/)?(.*)?)$/