+++ lib/puppet/type/wls_foreign_server.rb
@@ -43,6 +43,7 @@ module Puppet
     parameter :jmsmodule
     parameter :foreign_server_name
     parameter :password
+    parameter :timeout
     property :subdeployment
     property :defaulttargeting
     property :extraproperties