+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # History
 
+## 1.0.20
+
 ## 1.0.19
 - changed property and removed propertyvalues of wls_mail_session, now there is no change detected when there is a different sorting
 - changed users of wls_group, no change when there is a different sorting
 - removed extrapropertiesvalues of wls_datasource, you can use now extraproperties with key=value,key1=value
 - removed extrapropertiesvalues of wls_foreign_server, you can use now extraproperties with key=value,key1=value
+- New Order parameter on the wls_authentication_provider type so you can change the order
+- Timout parameter for all the wls types, can be used to override the default timeout (120s) on every wls resource
+- MaxMessageSize parameter for wls_server resource type
 
 ## 1.0.18
 - small fix for the server_template type with the arguments parameter